#555 Made UseExtendedTestCommandInspection a self-contained inspection
